Clarify CoC intent, close loopholes.
Due to the events of the last year or so, it's become clear that certain
extremists and trolls will attempt to shield themselves from criticism
by misapplying community guidelines and tolerance to forward their own
hateful and intolerant agenda. While I've always maintained an open mind
towards political affiliations and belief systems, it is with a heavy
heart that I've had to add language to explicitly indicate that I can't
put up with ideologies which explicitly go against the spirit of these
guidelines.

Nazi punks et al, fuck off.

And frankly I don't care what your ideology is supposed to be if you're
abusive and awful. <3