Revert to the purest form of Classpath exception.
The FYI paragraph at the end is being moved to the NOTICE where it belongs.

Eventually, I want to move to the full LGPL agreement but referencing the AGPL instead of the GPL. 

The only issue is that no one else has done that to my knowledge and it creates unknowns with possible confusion. The current strategy at least keeps things safe and straight forward. 

It works just like Java (ignoring Oracle's legal scuffles with Google) with the only difference being if you make changes you have to distribute them even if you're only running a service.

The other difference, that still needs to be noted is that I am willing to dual license it with a proprietary license. Both contributes and capitalist funding are helpful to the project.